feat(mcp): extract buildFocusManifestValidation into @loopover/engine for offline loopover_validate_config
Move the focus-manifest validation-result builder into @loopover/engine and
leave a re-export shim at src/services/focus-manifest-validation.ts, so the
local loopover_validate_config MCP tool computes the result in-process instead
of proxying to POST /v1/validate/focus-manifest. The unknownTopLevelWarnings
helper it depends on co-moves into the engine (single source of truth shared
with the self-host config linter). Bump packages/loopover-mcp's @loopover/engine
dependency to ^3.0.0 so it tracks the workspace engine, mirroring
packages/loopover-miner. Remote server behavior is unchanged via the shim.

Closes #6269
