+++ packages/loopover-mcp/package.json
@@ -38,7 +38,7 @@
     "build": "node --check bin/loopover-mcp.js && node --check lib/cli-error.js && node --check lib/local-branch.js && node --check lib/format-table.js && node --check scripts/gittensor-score-preview.mjs"
   },
   "dependencies": {
-    "@loopover/engine": "^1.0.0",
+    "@loopover/engine": "^3.0.0",
     "@modelcontextprotocol/sdk": "1.29.0",
     "zod": "^4.4.3"
   },
